完善

+++ src/pages/home/components/folder/store/actionCreators.js
@@ -16,6 +16,10 @@ const updateFolder = (folderData) => ({
 export const initialDeleteOneFiles = (file,folder,index) => {
   return async dispatch => {
     const result = await request.post('/api/deleteOneFile',{file,folder});
+    if(result.data.status === 0){
+      message.error(result.data.msg);
+      return;
+    }
     dispatch(deleteFile(index));
     message.success(result.data.msg);
   }
@@ -23,7 +27,7 @@ export const initialDeleteOneFiles = (file,folder,index) => {
 
 export const initialUpdateFolderData = (folderName) => {
   return async dispatch => {
-    const result = await request.get(`/api/openFolder?folderName=${folderName}`);
+    const result = await request.post('/api/openFolder',{folderName});
     dispatch(updateFolder(result.data.filesArr));
   }
 }
